refac : 사이드바 컴포넌트 분리
