+++ src/components/Sidebar/Sidebar.tsx
@@ -1,15 +1,18 @@
 import { Logo } from "@/assets/svg";
 import Close from "@/assets/svg/home/Close";
-import Text from "@/components/common/Text/Text";
 import LoginButton from "@/components/Header/LoginButton/LoginButton";
+import GuideItem from "@/components/Sidebar/Item/GuideItem";
+import MenuItem from "@/components/Sidebar/Item/MenuItem";
 import { Menus } from "@/core/Menu";
 import { useUser } from "@/hooks/useUser";
-import Icon from "@/pages/home/components/common/Icon";
-import { Drawer, Flex } from "antd";
-import { useLocation, useNavigate } from "react-router-dom";
+import { Drawer } from "antd";
 import styled from "styled-components";
 
-export default function Sidebar({ open, onClose }) {
+type SidebarProps = {
+    open: boolean;
+    onClose: () => void;
+};
+export default function Sidebar({ open, onClose }: SidebarProps) {
     const { userData } = useUser();
 
     return (
@@ -72,70 +75,3 @@ const BodyContainer = styled.div`
     height: 100%;
     padding: 12px 0;
 `;
-
-type MenuItemProps = {
-    menu: {
-        label: string;
-        path: string;
-    };
-    onClose: () => void;
-};
-const MenuItem = ({ menu, onClose }: MenuItemProps) => {
-    const location = useLocation();
-    const selected = location.pathname === menu.path;
-
-    const font = selected ? "b2_16_med" : "b2_16_reg";
-    const color = selected ? "primary" : "G_400";
-
-    const navigate = useNavigate();
-    const handleOnNavigate = () => {
-        navigate(menu.path);
-        onClose();
-    };
-    return (
-        <Item onClick={handleOnNavigate}>
-            {selected && <Rectangle />}
-            <Text font={font} color={color}>
-                {menu.label}
-            </Text>
-        </Item>
-    );
-};
-const Item = styled.div`
-    width: 100%;
-    padding: 14px 20px;
-    position: relative;
-    cursor: pointer;
-`;
-
-const Rectangle = styled.div`
-    position: absolute;
-    left: 0;
-    top: 0;
-    width: 4px;
-    height: 100%;
-    border-radius: 0px 2px 2px 0px;
-    background: ${({ theme }) => theme.colors.primary};
-`;
-
-const GuideItem = () => {
-    const handleOnGuide = () => {
-        window.open(
-            "https://pocket-prompt.notion.site/da477857a0cc44888b06dd23cf6682ff",
-            "_blank"
-        );
-    };
-    return (
-        <Item onClick={handleOnGuide}>
-            <Flex justify="space-between">
-                <Flex gap={8}>
-                    <Icon name="Book" />
-                    <Text font="b2_16_reg" color="primary">
-                        Guide
-                    </Text>
-                </Flex>
-                <Icon name="ExportSquare" />
-            </Flex>
-        </Item>
-    );
-};
